boot: prepare for rewrite of boot sysmodule
