+++ stratosphere/boot/boot.json
